+++ src/clld_corpus_plugin/datatables.py
@@ -1,6 +1,16 @@
+from clld.db.models.common import Contribution
+from clld.db.models.common import Language
+from clld.db.models.common import Parameter
+from clld.db.models.common import Sentence_files
+from clld.db.models.common import Value
+from clld.db.models.common import ValueSentence
+from clld.db.models.common import ValueSet
+from clld.web.datatables.base import Col
 from clld.web.datatables.base import DataTable
-from clld.web.datatables.base import LinkCol, Col
-from clld.web.datatables.sentence import Sentences, AudioCol
+from clld.web.datatables.base import LinkCol
+from clld.web.datatables.sentence import AudioCol
+from clld.web.datatables.sentence import Sentences as orig_Sentences
+from clld_corpus_plugin import models
 
 
 class CountCol(Col):
@@ -23,6 +33,16 @@ class Texts(DataTable):
 class Speakers(DataTable):
     def col_defs(self):
         return [Col(self, "id"), LinkCol(self, "name")]
+
+
+class Sentences(orig_Sentences):
+    __constraints__ = [Parameter, Contribution, Language]
+
+    def base_query(self, query):
+        query = super().base_query(query)
+        if self.contribution:
+            query = query.filter(models.RichSentence.contribution == self.contribution)
+        return query
 
 
 class SentencesWithAudio(Sentences):