+++ src/clld_corpus_plugin/models.py
@@ -1,10 +1,13 @@
+from clld import interfaces
 from clld.db.meta import Base
+from clld.db.meta import CustomModelMixin
 from clld.db.meta import PolymorphicBaseMixin
 from clld.db.models import IdNameDescriptionMixin
 from clld.db.models import Sentence
 from clld.db.models.common import Contribution
 from clld.db.models.common import HasSourceMixin
 from clld.db.models.common import Language
+from clld.interfaces import ISentence
 from sqlalchemy import JSON
 from sqlalchemy import Column
 from sqlalchemy import ForeignKey
@@ -12,10 +15,14 @@ from sqlalchemy import Integer
 from sqlalchemy import String
 from sqlalchemy import Unicode
 from sqlalchemy import UniqueConstraint
+from sqlalchemy import text
 from sqlalchemy.orm import relationship
 from zope.interface import implementer
-from clld_corpus_plugin.interfaces import IText, ITag, ISpeaker
-from clld_corpus_plugin.interfaces import IWordform, IMeaning
+from clld_corpus_plugin.interfaces import IMeaning
+from clld_corpus_plugin.interfaces import ISpeaker
+from clld_corpus_plugin.interfaces import ITag
+from clld_corpus_plugin.interfaces import IText
+from clld_corpus_plugin.interfaces import IWordform
 
 
 @implementer(ITag)
@@ -28,13 +35,21 @@ class Speaker(Base, IdNameDescriptionMixin):
     pass
 
 
+@implementer(interfaces.ISentence)
+class RichSentence(CustomModelMixin, Sentence):
+    pk = Column(Integer, ForeignKey("sentence.pk"), primary_key=True)
+
+    contribution_pk = Column(Integer, ForeignKey("contribution.pk"))
+    contribution = relationship(Contribution, backref="sentences")
+
+
 class SpeakerSentence(Base, PolymorphicBaseMixin):
     __table_args__ = (UniqueConstraint("speaker_pk", "sentence_pk"),)
 
     sentence_pk = Column(Integer, ForeignKey("sentence.pk"), nullable=False)
     speaker_pk = Column(Integer, ForeignKey("speaker.pk"), nullable=False)
     speaker = relationship(Speaker, innerjoin=True, backref="sentences")
-    sentence = relationship(Sentence, innerjoin=True, backref="speaker")
+    sentence = relationship(RichSentence, innerjoin=True, backref="speaker")
 
 
 @implementer(IText)
@@ -61,7 +76,7 @@ class SentenceTag(Base, PolymorphicBaseMixin):
 
     sentence_pk = Column(Integer, ForeignKey("sentence.pk"), nullable=False)
     tag_pk = Column(Integer, ForeignKey("tag.pk"), nullable=False)
-    sentence = relationship(Sentence, innerjoin=True, backref="tags")
+    sentence = relationship(RichSentence, innerjoin=True, backref="tags")
     tag = relationship(Tag, innerjoin=True, backref="sentences")
 
 
@@ -100,14 +115,17 @@ class TextSentence(Base, PolymorphicBaseMixin):
     phrase_number = Column(Integer, nullable=True)
 
     text = relationship(
-        Text, innerjoin=True, backref="sentences", order_by="desc(TextSentence.record_number)"
+        Text,
+        innerjoin=True,
+        backref="sentences",
+        order_by="desc(TextSentence.record_number)",
     )
-    sentence = relationship(Sentence, innerjoin=True, backref="text_assocs")
+    sentence = relationship(RichSentence, innerjoin=True, backref="text_assocs")
 
 
 class SentencePart(Base):
     form_pk = Column(Integer, ForeignKey("wordform.pk"))
     sentence_pk = Column(Integer, ForeignKey("sentence.pk"))
     form = relationship(Wordform, backref="sentence_assocs")
-    sentence = relationship(Sentence, backref="forms")
+    sentence = relationship(RichSentence, backref="forms")
     index = Column(Integer)