+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 * sentence parts have no meaning for now
 * slightly updated sentence detail view
+* sentences can belong to a contribution
 
 ## [0.0.7] - 2022-11-27